requirements.txt
keras-rl2=1.0.5
numpy=1.26.2
pip=23.2.1
setuptools=65.5.0
tensorflow=2.15.0
treys=0.1.8
